Base of WhiteBoard with login system and layout
Included in the initial commit is a functioning login system, with an
email verification, password reset and registration system. The layout
is a very rough design and can be changed if we see the need. There is
a lot of rough edges on the style of the pages, for example the log in
button. Another thing is the errors after a failed log in need to have
a better place to be displayed.