Courses Revamp
I changed the courses page to only show the sections a user is enrolled
in. You can click on the section id and you will be redirected to a
page about that course. There needs to be permissions on that page,
which will verify that the user is enrolled in that section. I also
added the search feature that will autocomplete suggestions. All of
this is very rough, but it will be a solid base. I am going to host
this branch at 54.200.146.119.